Always tick spawners in loaded chunks + Refactor of AnchorTask
When a user was inside a given chunk with a spawner, the spawner might not trigger because
of the activation range.
Anchors are supposed to ignore this limit - This change makes sure, spawners are always active.

The hack to not tick chunks with players in them is only in there because of the randomTicks done.
But this can still cause chunks to get double the randomTicks they are supposed to because
there is no good way (that I know of) to check when the server did them on a given chunk.